Update README.md
This project takes locally saved data about movies/movie ratings, and will generate recommendations using Euclidean distance. 

This was a school project intended to show some of my Python skills. You can use the program by downloading "data" files and saving them to the working directory for where you run code. The first thing the program will ask for is the name of the folder, the name of the movie data, the critics ratings, and finally the personal ratings. 

An example of what to enter in the command line upon executing the program would be: data IMDB.csv ratings.csv p2.csv
This will generate movie recommendations for the person whose movie preferences are saved in the p2.csv file.